refactor: modularize student dashboard for better maintainability
