+++ client/src/modules/dashboard/DashboardPage.jsx
@@ -41,6 +41,9 @@ import Button from "../../shared/components/Button";
 import Navbar from "../../shared/landing/Navbar";
 import { getAnalysisHistory, getSkillTrends } from "./services/dashboardService";
 import { getRecruiterJobs } from "../recruiter-jobs/services/jobPostingService";
+import SuggestionItem from "./components/SuggestionItem";
+import StatCard from "./components/StatCard";
+import PerformanceTrend from "./components/PerformanceTrend";
 
 const ROLE_LABELS = {
   student: "Student",
@@ -48,92 +51,6 @@ const ROLE_LABELS = {
   recruiter: "Recruiter",
 };
 
-const CustomTooltip = ({ active, payload, label }) => {
-  if (active && payload && payload.length) {
-    return (
-      <div className="bg-slate-900 border border-gray-200 dark:border-white/10 p-3 rounded-lg shadow-xl backdrop-blur-md">
-        <p className="text-xs text-slate-400 mb-1">{payload[0].payload.fullDate}</p>
-        <p className="text-sm font-bold text-blue-400">Score: {payload[0].value}%</p>
-      </div>
-    );
-  }
-  return null;
-};
-
-const SuggestionItem = ({ suggestion }) => {
-  // Handle both string and object formats for robustness
-  const text = typeof suggestion === "string" ? suggestion : (suggestion?.text || "");
-  const priority = suggestion?.priority || "";
-
-  const suggestionDetails = useMemo(() => {
-    const safeText = text.toLowerCase();
-    
-    if (priority === "Critical" || safeText.includes("urgent") || safeText.includes("missing")) {
-      return {
-        icon: <AlertCircle size={16} className="text-red-400" />,
-        bg: "bg-red-500/5",
-        border: "border-red-500/20",
-        label: "Critical Action",
-        textColor: "text-red-300"
-      };
-    }
-    
-    if (priority === "Strategic" || safeText.includes("add") || safeText.includes("integrate") || safeText.includes("projects")) {
-      return {
-        icon: <Sparkles size={16} className="text-blue-400" />,
-        bg: "bg-blue-500/5",
-        border: "border-blue-500/20",
-        label: "Strategic Tip",
-        textColor: "text-blue-300"
-      };
-    }
-
-    if (priority === "Optimization" || safeText.includes("metric") || safeText.includes("measure") || safeText.includes("highlight")) {
-      return {
-        icon: <TrendingUp size={16} className="text-emerald-400" />,
-        bg: "bg-emerald-500/5",
-        border: "border-emerald-500/20",
-        label: "Growth Area",
-        textColor: "text-emerald-300"
-      };
-    }
-
-    return {
-      icon: <Zap size={16} className="text-amber-400" />,
-      bg: "bg-amber-500/5",
-      border: "border-amber-500/20",
-      label: "Critical Action",
-      textColor: "text-amber-300"
-    };
-  }, [suggestion, text, priority]);
-
-  const { icon, bg, border, label, textColor } = suggestionDetails;
-
-  return (
-    <div className={`group flex flex-col gap-2 p-4 rounded-xl border ${bg} ${border} transition-all duration-300 hover:bg-white/[0.02]`}>
-      <div className="flex items-center justify-between">
-        <div className="flex items-center gap-2">
-          <div className={`p-1.5 rounded-lg ${bg} border border-white/5`}>
-            {icon}
-          </div>
-          <span className={`text-[10px] font-bold uppercase tracking-widest ${textColor}`}>{label}</span>
-        </div>
-        <ChevronRight size={14} className="text-slate-600 group-hover:translate-x-0.5 transition-transform" />
-      </div>
-      <p className="text-sm text-gray-700 dark:text-slate-300 leading-relaxed font-medium">
-        {text.split(' ').map((word, i) => {
-          const cleanWord = word.toLowerCase().replace(/[^a-z]/g, '');
-          const isHighlight = ['skills', 'projects', 'metrics', 'achievements', 'keywords', 'experience', 'ats', 'readability'].includes(cleanWord);
-          return (
-            <span key={i} className={isHighlight ? 'text-white font-bold' : ''}>
-              {word}{' '}
-            </span>
-          );
-        })}
-      </p>
-    </div>
-  );
-};
 
 const DashboardPage = () => {
   const dispatch = useDispatch();
@@ -262,40 +179,25 @@ const DashboardPage = () => {
 
         {/* User Stats Grid */}
         <section className="grid gap-4 sm:gap-6 md:grid-cols-3 grid-cols-1 sm:grid-cols-2">
-          <div className="group relative rounded-2xl border border-gray-200 dark:border-white/10 bg-gray-100 dark:bg-slate-900/50 p-5 shadow-2xl backdrop-blur-md transition-all hover:border-blue-500/30">
-            <div className="absolute inset-0 rounded-2xl bg-gradient-to-br from-blue-500/5 to-transparent opacity-0 group-hover:opacity-100 transition-opacity"></div>
-            <div className="relative">
-              <div className="mb-4 flex h-12 w-12 items-center justify-center rounded-xl bg-blue-500/20 text-blue-400 group-hover:scale-110 transition-transform">
-                <User size={24} />
-              </div>
-              <p className="text-xs font-semibold text-gray-500 dark:text-slate-500 uppercase tracking-widest">Profile Name</p>
-              <p className="mt-1 font-bold text-lg text-gray-900 dark:text-slate-100 break-words">{user?.name || "Not available"}</p>
-            </div>
-          </div>
-
-          <div className="group relative rounded-2xl border border-gray-200 dark:border-white/10 bg-gray-100 dark:bg-slate-900/50 p-5 shadow-2xl backdrop-blur-md transition-all hover:border-emerald-500/30">
-             <div className="absolute inset-0 rounded-2xl bg-gradient-to-br from-emerald-500/5 to-transparent opacity-0 group-hover:opacity-100 transition-opacity"></div>
-            <div className="relative">
-              <div className="mb-4 flex h-12 w-12 items-center justify-center rounded-xl bg-emerald-500/20 text-emerald-400 group-hover:scale-110 transition-transform">
-                <Mail size={24} />
-              </div>
-              <p className="text-xs font-semibold text-gray-500 dark:text-slate-500 uppercase tracking-widest">Email Address</p>
-              <p className="mt-1 font-bold text-lg text-gray-900 dark:text-slate-100 break-words">{user?.email || "Not available"}</p>
-            </div>
-          </div>
-
-          <div className="group relative rounded-2xl border border-gray-200 dark:border-white/10 bg-gray-100 dark:bg-slate-900/50 p-5 shadow-2xl backdrop-blur-md transition-all hover:border-violet-500/30 sm:col-span-2 md:col-span-1">
-             <div className="absolute inset-0 rounded-2xl bg-gradient-to-br from-violet-500/5 to-transparent opacity-0 group-hover:opacity-100 transition-opacity"></div>
-            <div className="relative">
-              <div className="mb-4 flex h-12 w-12 items-center justify-center rounded-xl bg-violet-500/20 text-violet-400 group-hover:scale-110 transition-transform">
-                <Shield size={24} />
-              </div>
-              <p className="text-xs font-semibold text-gray-500 dark:text-slate-500 uppercase tracking-widest">Access Role</p>
-              <p className="mt-1 font-bold text-lg text-gray-900 dark:text-slate-100">
-                {ROLE_LABELS[user?.role] || user?.role || "Not available"}
-              </p>
-            </div>
-          </div>
+          <StatCard 
+            icon={User} 
+            label="Profile Name" 
+            value={user?.name || "Not available"} 
+            color="blue" 
+          />
+          <StatCard 
+            icon={Mail} 
+            label="Email Address" 
+            value={user?.email || "Not available"} 
+            color="emerald" 
+          />
+          <StatCard 
+            icon={Shield} 
+            label="Access Role" 
+            value={ROLE_LABELS[user?.role] || user?.role || "Not available"} 
+            color="violet" 
+            className="sm:col-span-2 md:col-span-1"
+          />
         </section>
 
         {/* Recruiter Specific Stats Grid */}
@@ -345,64 +247,21 @@ const DashboardPage = () => {
             
             {/* Score Trend Chart - Student Only */}
             {isStudent && (
-              <div className="rounded-2xl border border-gray-200 dark:border-white/10 bg-gray-100 dark:bg-slate-900/50 overflow-hidden backdrop-blur-md">
-                <div className="border-b border-white/5 bg-white/5 px-6 py-4 flex items-center justify-between">
-                  <div className="flex items-center gap-2">
-                    <BarChart3 className="text-blue-400" size={20} />
-                    <h2 className="text-lg font-bold">Performance Trend</h2>
-                  </div>
-                  <div className="flex items-center gap-1.5 text-xs text-slate-400 bg-white/5 px-2 py-1 rounded-md">
-                    <Calendar size={12} />
-                    <span>Last {history.length} Analyses</span>
-                  </div>
-                </div>
-                <div className="p-6 h-[280px] min-h-[280px] w-full">
-                  {history.length > 1 ? (
-                    <ResponsiveContainer width="100%" height="100%">
-                      <AreaChart data={chartData}>
-                        <defs>
-                          <linearGradient id="colorScore" x1="0" y1="0" x2="0" y2="1">
-                            <stop offset="5%" stopColor="#3b82f6" stopOpacity={0.3}/>
-                            <stop offset="95%" stopColor="#3b82f6" stopOpacity={0}/>
-                          </linearGradient>
-                        </defs>
-                        <CartesianGrid strokeDasharray="3 3" stroke="#ffffff05" vertical={false} />
-                        <XAxis 
-                          dataKey="date" 
-                          stroke="#94a3b8" 
-                          fontSize={10} 
-                          tickLine={false} 
-                          axisLine={false}
-                          dy={10}
-                        />
-                        <YAxis 
-                          stroke="#94a3b8" 
-                          fontSize={10} 
-                          tickLine={false} 
-                          axisLine={false}
-                          domain={[0, 100]}
-                          tickFormatter={(val) => `${val}%`}
-                        />
-                        <RechartsTooltip content={<CustomTooltip />} />
-                        <Area 
-                          type="monotone" 
-                          dataKey="score" 
-                          stroke="#3b82f6" 
-                          strokeWidth={2}
-                          fillOpacity={1} 
-                          fill="url(#colorScore)" 
-                          animationDuration={1500}
-                        />
-                      </AreaChart>
-                    </ResponsiveContainer>
-                  ) : (
-                    <div className="h-full flex flex-col items-center justify-center text-gray-500 dark:text-slate-500">
-                      <Activity size={40} className="opacity-20 mb-3" />
-                      <p className="text-sm">Need at least 2 analyses to show trend</p>
-                    </div>
-                  )}
-                </div>
-              </div>
+              <PerformanceTrend 
+                data={chartData} 
+                historyLength={history.length} 
+                customTooltip={({ active, payload }) => {
+                  if (active && payload && payload.length) {
+                    return (
+                      <div className="bg-slate-900 border border-gray-200 dark:border-white/10 p-3 rounded-lg shadow-xl backdrop-blur-md">
+                        <p className="text-xs text-slate-400 mb-1">{payload[0].payload.fullDate}</p>
+                        <p className="text-sm font-bold text-blue-400">Score: {payload[0].value}%</p>
+                      </div>
+                    );
+                  }
+                  return null;
+                }}
+              />
             )}
 
             {/* Skill Trends - Student Only */}
@@ -585,7 +444,14 @@ const DashboardPage = () => {
                             <span className="absolute text-xl font-black">{latestAnalysis.score}%</span>
                           </div>
                           <div>
-                            <p className="text-xs font-bold text-gray-500 dark:text-slate-500 uppercase tracking-wider">Overall Score</p>
+                            <div className="flex items-center gap-2 mb-1">
+                              <p className="text-xs font-bold text-gray-500 dark:text-slate-500 uppercase tracking-wider">Overall Score</p>
+                              {latestAnalysis.mode === "benchmark" && (
+                                <span className="px-1.5 py-0.5 rounded bg-blue-500/10 text-blue-400 text-[8px] font-black uppercase tracking-tighter border border-blue-500/20">
+                                  Benchmark
+                                </span>
+                              )}
+                            </div>
                             <p className="text-2xl font-black text-white">{latestAnalysis.classification}</p>
                           </div>
                         </div>
@@ -673,6 +539,9 @@ const DashboardPage = () => {
                             <td className="px-6 py-4 text-sm font-bold whitespace-nowrap">
                               <span className={`${item.score >= 70 ? "text-emerald-400" : item.score >= 40 ? "text-yellow-400" : "text-red-400"}`}>
                                 {item.score}%
+                                {item.mode === "benchmark" && (
+                                  <span className="ml-2 text-[8px] opacity-60 text-blue-400 font-black uppercase">BM</span>
+                                )}
                               </span>
                             </td>
                             <td className="px-6 py-4 text-sm whitespace-nowrap">
